+++ src/gpu/ganesh/vk/GrVkGpu.cpp
@@ -28,13 +28,11 @@
 #include "src/gpu/ganesh/GrTexture.h"
 #include "src/gpu/ganesh/GrThreadSafePipelineBuilder.h"
 #include "src/gpu/ganesh/SkGr.h"
-#include "src/gpu/ganesh/vk/GrVkAMDMemoryAllocator.h"
 #include "src/gpu/ganesh/vk/GrVkBuffer.h"
 #include "src/gpu/ganesh/vk/GrVkCommandBuffer.h"
 #include "src/gpu/ganesh/vk/GrVkCommandPool.h"
 #include "src/gpu/ganesh/vk/GrVkFramebuffer.h"
 #include "src/gpu/ganesh/vk/GrVkImage.h"
-#include "src/gpu/ganesh/vk/GrVkInterface.h"
 #include "src/gpu/ganesh/vk/GrVkMemory.h"
 #include "src/gpu/ganesh/vk/GrVkOpsRenderPass.h"
 #include "src/gpu/ganesh/vk/GrVkPipeline.h"
@@ -44,6 +42,8 @@
 #include "src/gpu/ganesh/vk/GrVkSemaphore.h"
 #include "src/gpu/ganesh/vk/GrVkTexture.h"
 #include "src/gpu/ganesh/vk/GrVkTextureRenderTarget.h"
+#include "src/gpu/vk/VulkanAMDMemoryAllocator.h"
+#include "src/gpu/vk/VulkanInterface.h"
 #include "src/image/SkImage_Gpu.h"
 #include "src/image/SkSurface_Gpu.h"
 
@@ -101,15 +101,15 @@ sk_sp<GrGpu> GrVkGpu::Make(const GrVkBackendContext& backendContext,
     instanceVersion = std::min(instanceVersion, apiVersion);
     physDevVersion = std::min(physDevVersion, apiVersion);
 
-    sk_sp<const GrVkInterface> interface;
+    sk_sp<const skgpu::VulkanInterface> interface;
 
     if (backendContext.fVkExtensions) {
-        interface.reset(new GrVkInterface(backendContext.fGetProc,
-                                          backendContext.fInstance,
-                                          backendContext.fDevice,
-                                          instanceVersion,
-                                          physDevVersion,
-                                          backendContext.fVkExtensions));
+        interface.reset(new skgpu::VulkanInterface(backendContext.fGetProc,
+                                                   backendContext.fInstance,
+                                                   backendContext.fDevice,
+                                                   instanceVersion,
+                                                   physDevVersion,
+                                                   backendContext.fVkExtensions));
         if (!interface->validate(instanceVersion, physDevVersion, backendContext.fVkExtensions)) {
             return nullptr;
         }
@@ -123,12 +123,12 @@ sk_sp<GrGpu> GrVkGpu::Make(const GrVkBackendContext& backendContext,
             extensions.init(backendContext.fGetProc, backendContext.fInstance,
                             backendContext.fPhysicalDevice, 0, nullptr, 1, &swapChainExtName);
         }
-        interface.reset(new GrVkInterface(backendContext.fGetProc,
-                                          backendContext.fInstance,
-                                          backendContext.fDevice,
-                                          instanceVersion,
-                                          physDevVersion,
-                                          &extensions));
+        interface.reset(new skgpu::VulkanInterface(backendContext.fGetProc,
+                                                   backendContext.fInstance,
+                                                   backendContext.fDevice,
+                                                   instanceVersion,
+                                                   physDevVersion,
+                                                   &extensions));
         if (!interface->validate(instanceVersion, physDevVersion, &extensions)) {
             return nullptr;
         }
@@ -177,14 +177,17 @@ sk_sp<GrGpu> GrVkGpu::Make(const GrVkBackendContext& backendContext,
         return nullptr;
     }
 
-    sk_sp<GrVkMemoryAllocator> memoryAllocator = backendContext.fMemoryAllocator;
+    sk_sp<skgpu::VulkanMemoryAllocator> memoryAllocator = backendContext.fMemoryAllocator;
     if (!memoryAllocator) {
         // We were not given a memory allocator at creation
-        memoryAllocator = GrVkAMDMemoryAllocator::Make(backendContext.fInstance,
-                                                       backendContext.fPhysicalDevice,
-                                                       backendContext.fDevice, physDevVersion,
-                                                       backendContext.fVkExtensions, interface,
-                                                       caps.get());
+        bool mustUseCoherentHostVisibleMemory = caps->mustUseCoherentHostVisibleMemory();
+        memoryAllocator = skgpu::VulkanAMDMemoryAllocator::Make(backendContext.fInstance,
+                                                                backendContext.fPhysicalDevice,
+                                                                backendContext.fDevice,
+                                                                physDevVersion,
+                                                                backendContext.fVkExtensions,
+                                                                interface,
+                                                                mustUseCoherentHostVisibleMemory);
     }
     if (!memoryAllocator) {
         SkDEBUGFAIL("No supplied vulkan memory allocator and unable to create one internally.");
@@ -203,10 +206,13 @@ sk_sp<GrGpu> GrVkGpu::Make(const GrVkBackendContext& backendContext,
 
 ////////////////////////////////////////////////////////////////////////////////
 
-GrVkGpu::GrVkGpu(GrDirectContext* direct, const GrVkBackendContext& backendContext,
-                 sk_sp<GrVkCaps> caps, sk_sp<const GrVkInterface> interface,
-                 uint32_t instanceVersion, uint32_t physicalDeviceVersion,
-                 sk_sp<GrVkMemoryAllocator> memoryAllocator)
+GrVkGpu::GrVkGpu(GrDirectContext* direct,
+                 const GrVkBackendContext& backendContext,
+                 sk_sp<GrVkCaps> caps,
+                 sk_sp<const skgpu::VulkanInterface> interface,
+                 uint32_t instanceVersion,
+                 uint32_t physicalDeviceVersion,
+                 sk_sp<skgpu::VulkanMemoryAllocator> memoryAllocator)
         : INHERITED(direct)
         , fInterface(std::move(interface))
         , fMemoryAllocator(std::move(memoryAllocator))
@@ -835,14 +841,14 @@ bool GrVkGpu::uploadTexDataLinear(GrVkImage* texImage,
     };
     VkSubresourceLayout layout;
 
-    const GrVkInterface* interface = this->vkInterface();
+    const skgpu::VulkanInterface* interface = this->vkInterface();
 
     GR_VK_CALL(interface, GetImageSubresourceLayout(fDevice,
                                                     texImage->image(),
                                                     &subres,
                                                     &layout));
 
-    const GrVkAlloc& alloc = texImage->alloc();
+    const skgpu::VulkanAlloc& alloc = texImage->alloc();
     if (VK_NULL_HANDLE == alloc.fMemory) {
         return false;
     }