+++ src/gpu/graphite/Image_Graphite.h
@@ -19,6 +19,7 @@ class Recorder;
 
 class Image final : public SkImage_Base {
 public:
+    Image(uint32_t uniqueID, TextureProxyView, const SkColorInfo&);
     Image(TextureProxyView, const SkColorInfo&);
     ~Image() override;
 
@@ -63,6 +64,8 @@ public:
         return nullptr;
     }
 
+    TextureProxyView textureProxyView() const { return fTextureProxyView; }
+
 private:
 #if SK_SUPPORT_GPU
     std::unique_ptr<GrFragmentProcessor> onAsFragmentProcessor(
@@ -82,8 +85,7 @@ private:
     }
 #endif
 
-    std::tuple<TextureProxyView, SkColorType> onAsView(Recorder*,
-                                                       Mipmapped) const override;
+    sk_sp<SkImage> onMakeTextureImage(Recorder*, RequiredImageProperties) const override;
 
     TextureProxyView fTextureProxyView;
 };