+++ src/gpu/graphite/mtl/MtlResourceProvider.h
@@ -11,6 +11,7 @@
 #include "include/private/SkTHash.h"
 #include "src/gpu/graphite/DrawTypes.h"
 #include "src/gpu/graphite/ResourceProvider.h"
+#include "src/gpu/graphite/mtl/MtlGraphicsPipeline.h"
 
 #import <Metal/Metal.h>
 
@@ -26,9 +27,7 @@ public:
 
     sk_sp<Texture> createWrappedTexture(const BackendTexture&) override;
 
-    // Finds or creates a compatible DepthStencilState based on the enum
-    sk_cfp<id<MTLDepthStencilState>> findOrCreateCompatibleDepthStencilState(
-            const DepthStencilSettings&);
+    sk_sp<MtlGraphicsPipeline> findOrCreateLoadMSAAPipeline(const RenderPassDesc&);
 
 private:
     const MtlSharedContext* mtlSharedContext();
@@ -48,7 +47,11 @@ private:
     BackendTexture onCreateBackendTexture(SkISize dimensions, const TextureInfo&) override;
     void onDeleteBackendTexture(BackendTexture&) override;
 
+    sk_cfp<id<MTLDepthStencilState>> findOrCreateCompatibleDepthStencilState(
+            const DepthStencilSettings&);
+
     SkTHashMap<DepthStencilSettings, sk_cfp<id<MTLDepthStencilState>>> fDepthStencilStates;
+    SkTHashMap<uint64_t, sk_sp<MtlGraphicsPipeline>> fLoadMSAAPipelines;
 };
 
 } // namespace skgpu::graphite